Make sure we clear timeouts when we end
* Moved business logic for ending session to controller
* Fixed logic for setting timeouts
