+++ router/api/session.js
@@ -30,21 +30,7 @@ module.exports = function (router) {
       }
     )
   })
-  function addSession (user, session) {
-    User.update({ _id: user._id },
-      { $addToSet: { pastSessions: session._id } },
-      function (err, results) {
-        if (err) {
-          throw err
-        } else {
-          // print out what session was added to which user
-          if (results.nModified === 1) {
-            console.log(`${session._id} session was added to ` +
-            `${user._id}'s pastSessions`)
-          }
-        }
-      })
-  }
+  
   router.route('/session/end').post(function (req, res) {
     var data = req.body || {}
     var sessionId = data.sessionId
@@ -58,15 +44,13 @@ module.exports = function (router) {
         } else if (!session) {
           res.json({ err: 'No session found' })
         } else {
-          var student = session.student
-          var volunteer = session.volunteer
-          // add session to the student and volunteer's pastSessions
-          addSession(student, session)
-          if (volunteer) {
-            addSession(volunteer, session)
-          }
-          session.endSession()
-          res.json({ sessionId: session._id })
+          SessionCtrl.end(session, function (err, session) {
+            if (err) {
+              res.json({ err: err })
+            } else {
+              res.json({ sessionId: session._id })
+            }
+          })
         }
       }
     )
